[td] Move commit to files
