+++ mage_ai/frontend/components/VersionControl/Commit/index.tsx
@@ -1,4 +1,3 @@
-import NextLink from 'next/link';
 import { useEffect, useMemo, useState } from 'react';
 import { useMutation } from 'react-query';
 
@@ -27,7 +26,7 @@ import {
   UNITS_BETWEEN_ITEMS_IN_SECTIONS,
   UNITS_BETWEEN_SECTIONS,
 } from '@oracle/styles/units/spacing';
-import { capitalizeRemoveUnderscoreLower, pluralize } from '@utils/string';
+import { capitalizeRemoveUnderscoreLower } from '@utils/string';
 import { onSuccess } from '@api/utils/response';
 import { unique } from '@utils/array';
 
@@ -40,7 +39,7 @@ type CommitProps = {
     [fullPath: string]: boolean;
   };
   remotes: GitRemoteType[];
-  setSelectedFilePath: (filePath: string) => void;
+
   showError: (opts: any) => void;
   stagedFiles: {
     [fullPath: string]: boolean;
@@ -50,81 +49,16 @@ type CommitProps = {
 function Commit({
   branch,
   branches,
-  fetchBranch: fetchBranchProp,
   loading,
-  modifiedFiles,
   remotes,
-  setSelectedFilePath,
   showError,
-  stagedFiles,
 }: CommitProps) {
   const [actionBranchName, setActionBranchName] = useState<string>(branch?.name || '');
   const [actionError, setActionError] = useState<string>(null);
   const [actionProgress, setActionProgress] = useState<string>(null);
   const [actionRemoteName, setActionRemoteName] = useState<string>('');
-  const [commitMessage, setCommitMessage] = useState<string>('');
   const [pullRequest, setPullRequest] = useState<PullRequestType>({});
   const [repositoryName, setRepositoryName] = useState<string>(null);
-
-  const { data: dataBranch, mutate: fetchBranch } = api.git_branches.detail('with_logs', {
-    '_format': 'with_logs',
-  });
-  const logs = useMemo(() => dataBranch?.git_branch?.logs || [], [dataBranch]);
-
-  const logsMemo = useMemo(() => (
-    <Table
-      columnFlex={[1, 1, 1]}
-      columns={[
-        {
-          uuid: 'Author',
-        },
-        {
-          uuid: 'Date',
-        },
-        {
-          uuid: 'Message',
-        },
-      ]}
-      rows={logs.map(({
-        author,
-        date,
-        message,
-      }) => [
-        <Text default key="author" monospace small>
-          {author?.name}
-        </Text>,
-        <Text default key="date" monospace small>
-          {date}
-        </Text>,
-        <Text default key="message" monospace small>
-          {message}
-        </Text>,
-      ])}
-      uuid="git-branch-logs"
-    />
-  ), [logs]);
-
-  const stagedFilesPaths = useMemo(() => Object.keys(stagedFiles || {}), [stagedFiles]);
-  const stagedFilesCount = useMemo(() => stagedFilesPaths.length, [stagedFilesPaths]);
-
-  const [updateGitBranch, { isLoading: isLoadingUpdate }] = useMutation(
-    api.git_branches.useUpdate(branch?.name),
-    {
-      onSuccess: (response: any) => onSuccess(
-        response, {
-          callback: () => {
-            fetchBranch();
-            fetchBranchProp();
-            setCommitMessage('');
-          },
-          onErrorCallback: (response, errors) => showError({
-            errors,
-            response,
-          }),
-        },
-      ),
-    },
-  );
 
   const [actionGitBranch, { isLoading: isLoadingAction }] = useMutation(
     api.git_branches.useUpdate(branch?.name),
@@ -276,117 +210,6 @@ function Commit({
 
   return (
     <>
-      <Spacing mb={UNITS_BETWEEN_SECTIONS}>
-        <Spacing mb={PADDING_UNITS}>
-          <Accordion>
-            <AccordionPanel
-              noPaddingContent
-              title={stagedFilesCount >= 1 ? `Staged files (${stagedFilesCount})` : 'No staged files'}
-            >
-              {stagedFilesPaths?.map((filePath: string) => (
-                <Spacing key={filePath} my={1} px={PADDING_UNITS}>
-                  <FlexContainer justifyContent="space-between">
-                    <Link
-                      default
-                      monospace
-                      // @ts-ignore
-                      onClick={() => setSelectedFilePath(prev => prev === filePath ? null : filePath)}
-                      warning={modifiedFiles?.[filePath]}
-                    >
-                      {filePath}
-                    </Link>
-
-                    <Spacing mr={1} />
-
-                    {modifiedFiles?.[filePath] && (
-                      <Text warning>
-                        Modified after <NextLink
-                          href={`/version-control?tab=${TAB_FILES.uuid}`}
-                          passHref
-                        >
-                          <Link
-                            underline
-                            warning
-                          >
-                            staging
-                          </Link>
-                        </NextLink>
-                      </Text>
-                    )}
-                  </FlexContainer>
-                </Spacing>
-              ))}
-            </AccordionPanel>
-          </Accordion>
-        </Spacing>
-
-        <TextArea
-          disabled={stagedFilesCount === 0}
-          label="Commit message"
-          monospace
-          onChange={e => setCommitMessage(e.target.value)}
-          value={commitMessage || ''}
-        />
-
-        <Spacing mt={PADDING_UNITS}>
-          <FlexContainer alignItems="center">
-            <Button
-              disabled={stagedFilesCount === 0 || !(commitMessage?.length >= 1)}
-              loading={isLoadingUpdate}
-              onClick={() => {
-                // @ts-ignore
-                updateGitBranch({
-                  git_branch: {
-                    action_type: 'commit',
-                    files: stagedFilesPaths,
-                    message: commitMessage,
-                  },
-                });
-              }}
-              primary
-            >
-              Commit {pluralize('file', stagedFilesCount, true)} with message
-            </Button>
-
-            {stagedFilesCount === 0 && (
-              <>
-                <Spacing mr={1} />
-                <Text danger small>
-                  Please <NextLink
-                    href={`/version-control?tab=${TAB_FILES.uuid}`}
-                    passHref
-                  >
-                    <Link
-                      danger
-                      small
-                      underline
-                    >
-                      stage
-                    </Link>
-                  </NextLink> at least 1 file before committing.
-                </Text>
-              </>
-            )}
-          </FlexContainer>
-        </Spacing>
-      </Spacing>
-
-      <Spacing mb={UNITS_BETWEEN_SECTIONS}>
-        <Accordion>
-          <AccordionPanel
-            noPaddingContent
-            title="Logs"
-          >
-            {!dataBranch && (
-              <Spacing p={PADDING_UNITS}>
-                <Spinner inverted />
-              </Spacing>
-            )}
-            {dataBranch && logsMemo}
-          </AccordionPanel>
-        </Accordion>
-      </Spacing>
-
       <Spacing mb={UNITS_BETWEEN_SECTIONS}>
         <Spacing mb={1}>
           <Headline>
