+++ mage_ai/frontend/components/VersionControl/GitFiles/index.tsx
@@ -1,7 +1,10 @@
+import NextLink from 'next/link';
 import { useCallback, useMemo, useState } from 'react';
 import { useMutation } from 'react-query';
 import { useRouter } from 'next/router';
 
+import Accordion from '@oracle/components/Accordion';
+import AccordionPanel from '@oracle/components/Accordion/AccordionPanel';
 import Button from '@oracle/elements/Button';
 import Checkbox from '@oracle/elements/Checkbox';
 import Divider from '@oracle/elements/Divider';
@@ -11,7 +14,10 @@ import GitBranchType from '@interfaces/GitBranchType';
 import Headline from '@oracle/elements/Headline';
 import Link from '@oracle/elements/Link';
 import Spacing from '@oracle/elements/Spacing';
+import Spinner from '@oracle/components/Spinner';
+import Table from '@components/shared/Table';
 import Text from '@oracle/elements/Text';
+import TextArea from '@oracle/elements/Inputs/TextArea';
 import api from '@api';
 import { PADDING_UNITS, UNITS_BETWEEN_SECTIONS } from '@oracle/styles/units/spacing';
 import { PaginateArrowLeft, PaginateArrowRight } from '@oracle/icons';
@@ -22,6 +28,7 @@ import {
 } from '../constants';
 import { isEmptyObject } from '@utils/hash';
 import { onSuccess } from '@api/utils/response';
+import { pluralize } from '@utils/string';
 
 type GitFilesProps = {
   branch: GitBranchType;
@@ -29,6 +36,7 @@ type GitFilesProps = {
   modifiedFiles: {
     [fullPath: string]: boolean;
   };
+  setSelectedFilePath: (filePath: string) => void;
   showError: (opts: any) => void;
   stagedFiles: {
     [fullPath: string]: boolean;
@@ -40,7 +48,7 @@ type GitFilesProps = {
 
 function GitFiles({
   branch,
-  fetchBranch,
+  fetchBranch: fetchBranchProp,
   modifiedFiles,
   showError,
   stagedFiles,
@@ -48,6 +56,7 @@ function GitFiles({
 }: GitFilesProps) {
   const router = useRouter();
 
+  const [commitMessage, setCommitMessage] = useState<string>('');
   const [selectedFilesA, setSelectedFilesA] = useState<{
     [fullPath: string]: boolean;
   }>({});
@@ -61,10 +70,8 @@ function GitFiles({
       untrackedFiles,
     ]);
 
-  const stagedFilePaths: string[] =
-    useMemo(() => Object.keys(stagedFiles), [
-      stagedFiles,
-    ]);
+  const stagedFilePaths = useMemo(() => Object.keys(stagedFiles || {}), [stagedFiles]);
+  const stagedFilesCount = useMemo(() => stagedFilePaths.length, [stagedFilePaths]);
 
   const allFilesASelected =
     useMemo(() => Object.keys(selectedFilesA).length === unstagedFilePaths?.length, [
@@ -88,13 +95,13 @@ function GitFiles({
     onSuccess: (response: any) => onSuccess(
       response, {
         callback: () => {
-          fetchBranch();
+          fetchBranchProp();
           setSelectedFilesA({});
         },
         ...sharedUpdateProps,
       },
     ),
-  }), [fetchBranch, sharedUpdateProps]);
+  }), [fetchBranchProp, sharedUpdateProps]);
   const updateEndpoint = useMemo(() => api.git_branches.useUpdate(branch?.name), [branch]);
 
   const [updateGitBranch, { isLoading: isLoadingUpdate }] = useMutation(
@@ -111,7 +118,7 @@ function GitFiles({
       onSuccess: (response: any) => onSuccess(
         response, {
           callback: () => {
-            fetchBranch();
+            fetchBranchProp();
             setSelectedFilesB({});
           },
           ...sharedUpdateProps,
@@ -204,6 +211,63 @@ function GitFiles({
   }, []);
 
   const noFilesASelected: boolean = useMemo(() => isEmptyObject(selectedFilesA), [selectedFilesA]);
+
+  const { data: dataBranch, mutate: fetchBranch } = api.git_branches.detail('with_logs', {
+    '_format': 'with_logs',
+  });
+  const logs = useMemo(() => dataBranch?.git_branch?.logs || [], [dataBranch]);
+
+  const [updateGitBranchCommit, { isLoading: isLoadingUpdateCommit }] = useMutation(
+    api.git_branches.useUpdate(branch?.name),
+    {
+      onSuccess: (response: any) => onSuccess(
+        response, {
+          callback: () => {
+            fetchBranch();
+            fetchBranchProp();
+            setCommitMessage('');
+          },
+          onErrorCallback: (response, errors) => showError({
+            errors,
+            response,
+          }),
+        },
+      ),
+    },
+  );
+
+  const logsMemo = useMemo(() => (
+    <Table
+      columnFlex={[1, 1, 1]}
+      columns={[
+        {
+          uuid: 'Author',
+        },
+        {
+          uuid: 'Date',
+        },
+        {
+          uuid: 'Message',
+        },
+      ]}
+      rows={logs.map(({
+        author,
+        date,
+        message,
+      }) => [
+        <Text default key="author" monospace small>
+          {author?.name}
+        </Text>,
+        <Text default key="date" monospace small>
+          {date}
+        </Text>,
+        <Text default key="message" monospace small>
+          {message}
+        </Text>,
+      ])}
+      uuid="git-branch-logs"
+    />
+  ), [logs]);
 
   return (
     <>
@@ -321,6 +385,69 @@ function GitFiles({
             )}
           </Flex>
         </FlexContainer>
+      </Spacing>
+
+      <Spacing mb={UNITS_BETWEEN_SECTIONS}>
+        <Spacing mb={1}>
+          <Headline>
+            Commit
+          </Headline>
+        </Spacing>
+
+        <TextArea
+          disabled={stagedFilesCount === 0}
+          label="Commit message"
+          monospace
+          onChange={e => setCommitMessage(e.target.value)}
+          value={commitMessage || ''}
+        />
+
+        <Spacing mt={PADDING_UNITS}>
+          <FlexContainer alignItems="center">
+            <Button
+              disabled={stagedFilesCount === 0 || !(commitMessage?.length >= 1)}
+              loading={isLoadingUpdateCommit}
+              onClick={() => {
+                // @ts-ignore
+                updateGitBranchCommit({
+                  git_branch: {
+                    action_type: 'commit',
+                    files: stagedFilePaths,
+                    message: commitMessage,
+                  },
+                });
+              }}
+              primary
+            >
+              Commit {pluralize('file', stagedFilesCount, true)} with message
+            </Button>
+
+            {stagedFilesCount === 0 && (
+              <>
+                <Spacing mr={1} />
+                <Text danger small>
+                  Please stage at least 1 file before committing.
+                </Text>
+              </>
+            )}
+          </FlexContainer>
+        </Spacing>
+      </Spacing>
+
+      <Spacing mb={UNITS_BETWEEN_SECTIONS}>
+        <Accordion>
+          <AccordionPanel
+            noPaddingContent
+            title="Logs"
+          >
+            {!dataBranch && (
+              <Spacing p={PADDING_UNITS}>
+                <Spinner inverted />
+              </Spacing>
+            )}
+            {dataBranch && logsMemo}
+          </AccordionPanel>
+        </Accordion>
       </Spacing>
 
       <Spacing mb={UNITS_BETWEEN_SECTIONS}>
